requirements.txt
pyyaml
Pillow
numpy
tqdm
numba
matplotlib